Updated JSP Pages, mostly addEvents.jsp
addEvents still needs to get the radio buttons and checkboxes added for
recurring/which days in both the HTML and the javascript sections.

viewCalendar needs to be implemented (and the deleting from that page)

My Account Settings page needs to be done. Wasn't too sure what needed
to go there; we can use a similar check for logged-in users and
displaying the menu for showing team leader options most likely.